pyproject.toml
[project]
name = "my-doan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "opencv-python>=4.12.0.88",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "supervision>=0.26.1",
    "ultralytics>=8.3.228",
    "uvicorn[standard]>=0.38.0",
]
